requirements.txt
playwright
playwright-stealth
pydantic>=2.0
pydantic-settings
pyyaml
httpx
tenacity
apscheduler>=3.10,<4
structlog
python-dotenv
rich
click

# dev
pytest
pytest-asyncio
pytest-playwright
ruff
